requirements.txt
# Data generation
pandas==2.0.3
numpy==1.24.3
faker==18.13.0

# GCP integration
google-cloud-storage==2.10.0
google-cloud-bigquery==3.12.0

# Pipeline (install separately or in Dataflow template)
# apache-beam[gcp]==2.52.0
